+++ zarrs/src/array/codec/array_to_array/squeeze/squeeze_partial_decoder.rs
@@ -1,10 +1,13 @@
 use std::{num::NonZeroU64, sync::Arc};
 
-use itertools::izip;
+use itertools::{izip, Itertools};
 
-use crate::array::{
-    codec::{ArrayBytes, ArrayPartialDecoderTraits, ArraySubset, CodecError, CodecOptions},
-    ChunkRepresentation, DataType,
+use crate::{
+    array::{
+        codec::{ArrayBytes, ArrayPartialDecoderTraits, ArraySubset, CodecError, CodecOptions},
+        ArrayIndices, ChunkRepresentation, DataType,
+    },
+    indexer::{IncompatibleIndexerError, Indexer},
 };
 
 #[cfg(feature = "async")]
@@ -29,15 +32,16 @@ impl SqueezePartialDecoder {
     }
 }
 
-fn get_decoded_regions_squeezed(
+fn get_squeezed_array_subset(
     decoded_region: &ArraySubset,
     shape: &[NonZeroU64],
 ) -> Result<ArraySubset, CodecError> {
     if decoded_region.dimensionality() != shape.len() {
-        return Err(CodecError::InvalidArraySubsetDimensionalityError(
-            decoded_region.clone(),
+        return Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+            decoded_region.dimensionality(),
             shape.len(),
-        ));
+        )
+        .into());
     }
 
     let ranges = izip!(
@@ -52,6 +56,33 @@ fn get_decoded_regions_squeezed(
     Ok(decoded_region_squeeze)
 }
 
+fn get_squeezed_indexer(
+    indexer: &dyn Indexer,
+    shape: &[NonZeroU64],
+) -> Result<impl Indexer, CodecError> {
+    let indices = indexer
+        .iter_indices()
+        .map(|indices| {
+            if indices.len() == shape.len() {
+                Ok(indices
+                    .into_iter()
+                    .zip(shape)
+                    .filter_map(
+                        |(indices, &shape)| if shape.get() > 1 { Some(indices) } else { None },
+                    )
+                    .collect_vec())
+            } else {
+                Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+                    indices.len(),
+                    shape.len(),
+                ))
+            }
+        })
+        .collect::<Result<Vec<ArrayIndices>, _>>()?;
+
+    Ok(indices)
+}
+
 impl ArrayPartialDecoderTraits for SqueezePartialDecoder {
     fn data_type(&self) -> &DataType {
         self.decoded_representation.data_type()
@@ -63,18 +94,19 @@ impl ArrayPartialDecoderTraits for SqueezePartialDecoder {
 
     fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
-        // let Some(decoded_region) = indexer.as_array_subset() else {
-        //     todo!("Generic indexer support")
-        // };
-        let decoded_region = indexer;
-
-        let decoded_region_squeezed =
-            get_decoded_regions_squeezed(decoded_region, self.decoded_representation.shape())?;
-        self.input_handle
-            .partial_decode(&decoded_region_squeezed, options)
+        if let Some(array_subset) = indexer.as_array_subset() {
+            let array_subset_squeezed =
+                get_squeezed_array_subset(array_subset, self.decoded_representation.shape())?;
+            self.input_handle
+                .partial_decode(&array_subset_squeezed, options)
+        } else {
+            let indexer_squeezed =
+                get_squeezed_indexer(indexer, self.decoded_representation.shape())?;
+            self.input_handle.partial_decode(&indexer_squeezed, options)
+        }
     }
 }
 
@@ -108,18 +140,21 @@ impl AsyncArrayPartialDecoderTraits for AsyncSqueezePartialDecoder {
 
     async fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
-        // let Some(decoded_region) = indexer.as_array_subset() else {
-        //     todo!("Generic indexer support")
-        // };
-        let decoded_region = indexer;
-
-        let decoded_region_squeezed =
-            get_decoded_regions_squeezed(decoded_region, self.decoded_representation.shape())?;
-        self.input_handle
-            .partial_decode(&decoded_region_squeezed, options)
-            .await
+        if let Some(array_subset) = indexer.as_array_subset() {
+            let array_subset_squeezed =
+                get_squeezed_array_subset(array_subset, self.decoded_representation.shape())?;
+            self.input_handle
+                .partial_decode(&array_subset_squeezed, options)
+                .await
+        } else {
+            let indexer_squeezed =
+                get_squeezed_indexer(indexer, self.decoded_representation.shape())?;
+            self.input_handle
+                .partial_decode(&indexer_squeezed, options)
+                .await
+        }
     }
 }