+++ zarrs/src/array/codec/array_to_bytes/bytes/bytes_partial_decoder.rs
@@ -7,7 +7,7 @@ use crate::{
         codec::{ArrayPartialDecoderTraits, BytesPartialDecoderTraits, CodecError, CodecOptions},
         ArrayBytes, ArraySize, ChunkRepresentation, DataType,
     },
-    array_subset::{ArraySubset, IncompatibleArraySubsetAndShapeError},
+    indexer::IncompatibleIndexerError,
 };
 
 #[cfg(feature = "async")]
@@ -48,7 +48,7 @@ impl ArrayPartialDecoderTraits for BytesPartialDecoder {
 
     fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
         let Some(data_type_size) = self.data_type().fixed_size() else {
@@ -58,26 +58,27 @@ impl ArrayPartialDecoderTraits for BytesPartialDecoder {
             ));
         };
 
+        if indexer.dimensionality() != self.decoded_representation.dimensionality() {
+            return Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+                indexer.dimensionality(),
+                self.decoded_representation.dimensionality(),
+            )
+            .into());
+        }
+
         let chunk_shape = self.decoded_representation.shape_u64();
         // Get byte ranges
-        let byte_ranges = indexer
-            .byte_ranges(&chunk_shape, data_type_size)
-            .map_err(|_| {
-                IncompatibleArraySubsetAndShapeError::from((
-                    indexer.clone(),
-                    self.decoded_representation.shape_u64(),
-                ))
-            })?;
+        let mut byte_ranges = indexer.byte_ranges(&chunk_shape, data_type_size)?;
 
         // Decode
         let decoded = self
             .input_handle
-            .partial_decode_concat(&byte_ranges, options)?
+            .partial_decode_concat(&mut byte_ranges, options)?
             .map_or_else(
                 || {
                     let array_size = ArraySize::new(
                         self.decoded_representation.data_type().size(),
-                        indexer.num_elements(),
+                        indexer.len(),
                     );
                     ArrayBytes::new_fill_value(array_size, self.decoded_representation.fill_value())
                 },
@@ -131,7 +132,7 @@ impl AsyncArrayPartialDecoderTraits for AsyncBytesPartialDecoder {
 
     async fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
         let Some(data_type_size) = self.data_type().fixed_size() else {
@@ -142,34 +143,28 @@ impl AsyncArrayPartialDecoderTraits for AsyncBytesPartialDecoder {
         };
 
         if indexer.dimensionality() != self.decoded_representation.dimensionality() {
-            return Err(CodecError::InvalidArraySubsetDimensionalityError(
-                indexer.clone(),
+            return Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+                indexer.dimensionality(),
                 self.decoded_representation.dimensionality(),
-            ));
+            )
+            .into());
         }
 
         let chunk_shape = self.decoded_representation.shape_u64();
 
         // Get byte ranges
-        let byte_ranges = indexer
-            .byte_ranges(&chunk_shape, data_type_size)
-            .map_err(|_| {
-                IncompatibleArraySubsetAndShapeError::from((
-                    indexer.clone(),
-                    self.decoded_representation.shape_u64(),
-                ))
-            })?;
+        let mut byte_ranges = indexer.byte_ranges(&chunk_shape, data_type_size)?;
 
         // Decode
         let decoded = self
             .input_handle
-            .partial_decode_concat(&byte_ranges, options)
+            .partial_decode_concat(&mut byte_ranges, options)
             .await?
             .map_or_else(
                 || {
                     let array_size = ArraySize::new(
                         self.decoded_representation.data_type().size(),
-                        indexer.num_elements(),
+                        indexer.len(),
                     );
                     ArrayBytes::new_fill_value(array_size, self.decoded_representation.fill_value())
                 },